backend/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blis==1.3.3
build==1.4.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.3
chromadb==1.5.0
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
datasets==2.19.1
dill==0.3.8
distro==1.9.0
durationpy==0.10
einops==0.8.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
et_xmlfile==2.0.0
evaluate==0.4.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.3.1
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==1.2.9
langchain-core==1.2.9
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mmh3==5.2.0
monai==1.3.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.15
neo4j==6.1.0
networkx==3.5
nibabel==5.2.1
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.24.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pandas==2.3.3
peft==0.8.2
pillow==11.3.0
posthog==5.4.0
preshed==3.0.12
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pybase64==1.4.3
pycocotools==2.0.7
pydantic==2.12.5
pydantic_core==2.41.5
pydicom==3.0.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
PyPika==0.51.1
pyproject_hooks==1.2.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qimage2ndarray==1.10.0
QtAwesome==1.4.0
QtPy==2.4.3
referencing==0.37.0
regex==2025.9.18
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rich==14.3.2
rouge==1.0.1
rpds-py==0.30.0
safetensors==0.6.2
scikit-learn==1.8.0
scipy==1.17.0
sentencepiece==0.2.1
shellingham==1.5.4
simple-slice-viewer==0.97
simpleitk==2.5.2
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.66.2
transformers==4.44.2
tweepy==4.14.0
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0